+++ src/azure-cli/azure/cli/command_modules/vm/tests/hybrid_2018_03_01/test_vm_image.py
@@ -66,8 +66,9 @@ class TestVMImage(unittest.TestCase):
         # assert
         cli_ctx = DummyCli()
         cli_ctx.cloud = mock_cloud
-        with self.assertRaises(CLIError):
-            load_images_from_aliases_doc(cli_ctx)
+        images = load_images_from_aliases_doc(cli_ctx)
+        self.assertEqual(images[0], {'urnAlias': 'CentOS', 'publisher': 'OpenLogic',
+                                     'offer': 'CentOS', 'sku': '7.5', 'version': 'latest'})
 
 
 if __name__ == '__main__':